--- src/main/java/ladder/game/Connectable.java
@@ -1,6 +1,0 @@
-package ladder.game;
-
-@FunctionalInterface
-public interface Connectable {
-    boolean create();
-}
