+++ src/main/java/ladder/game/Participant.java
@@ -3,6 +3,7 @@ package ladder.game;
 public class Participant {
     private static final int PARTICIPANT_NAME_MAX_LENGTH = 5;
     private static final String PARTICIPANT_NAME_MAX_LENGTH_EXCEPTION = "참여자의 이름은 %d글자를 넘을 수 없습니다";
+
     private final String name;
 
     public Participant(String participant) {
