+++ src/main/java/ladder/structure/connection/Connectable.java
@@ -1,0 +1,6 @@
+package ladder.structure.connection;
+
+@FunctionalInterface
+public interface Connectable {
+    boolean connect();
+}
