+++ src/main/java/ladder/structure/connection/ConnectionStrategy.java
@@ -1,4 +1,4 @@
-package ladder.game;
+package ladder.structure.connection;
 
 @FunctionalInterface
 public interface ConnectionStrategy {
