+++ src/main/java/ladder/structure/connection/NoneConnection.java
@@ -1,4 +1,4 @@
-package ladder.game;
+package ladder.structure.connection;
 
 public class NoneConnection implements ConnectionStrategy {
     @Override
