+++ src/main/java/ladder/structure/connection/RandomConnection.java
@@ -1,11 +1,12 @@
-package ladder.game;
+package ladder.structure.connection;
 
 import java.util.Random;
 
 public class RandomConnection implements ConnectionStrategy {
     private static final Random random = new Random();
+
     private static boolean random(Connectable connection) {
-        return connection.create();
+        return connection.connect();
     }
 
     @Override
