+++ src/test/java/ladder/game/LadderTest.java
@@ -1,5 +1,7 @@
 package ladder.game;
 
+import ladder.structure.Ladder;
+import ladder.structure.connection.RandomConnection;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
